requirements.txt
numpy==1.26.4 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-osc==1.8.3 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0"
